% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-qc-checks.R
\name{append_flag_reason}
\alias{append_flag_reason}
\title{Helper function to append flag reasons}
\usage{
append_flag_reason(flags, idx, new_reason)
}
\description{
Helper function to append flag reasons
}
