% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-qc-checks.R
\name{check_condition_factor}
\alias{check_condition_factor}
\title{Check condition factor (K = weight / length^3 * 10^5)}
\usage{
check_condition_factor(data)
}
\description{
Check condition factor (K = weight / length^3 * 10^5)
}
