% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-qc-checks.R
\name{check_interobserver_agreement}
\alias{check_interobserver_agreement}
\title{Check inter-observer agreement for duplicate counts}
\usage{
check_interobserver_agreement(data)
}
\description{
Check inter-observer agreement for duplicate counts
}
