% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-qc-checks.R
\name{check_rate_of_change}
\alias{check_rate_of_change}
\title{Check rate of change for sudden spikes/drops}
\usage{
check_rate_of_change(data, max_change)
}
\description{
Check rate of change for sudden spikes/drops
}
