% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-qc-checks.R
\name{generate_qc_report}
\alias{generate_qc_report}
\title{Generate QC Report}
\usage{
generate_qc_report(qc_results, output_file = NULL)
}
\arguments{
\item{qc_results}{List output from run_qc_checks()}

\item{output_file}{Path to save report (optional)}
}
\description{
Generate QC Report
}
