% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-clean-checks.R
\name{log_info}
\alias{log_info}
\title{Log info message}
\usage{
log_info(..., verbose = F)
}
\arguments{
\item{...}{Message components passed to message()}

\item{verbose}{Logical, if TRUE the message is printed (default: FALSE)}
}
\value{
NULL invisibly
}
\description{
Prints a message to the console only when verbose is TRUE
}
