% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-eml-edi.R
\name{make_eml_edi}
\alias{make_eml_edi}
\title{Make metadata XML required for EDI data publication}
\usage{
make_eml_edi(
  data_file_names,
  attributes_file_names,
  title,
  maintenance,
  edi_number
)
}
\arguments{
\item{data_file_names}{A list of file names for data tables to be included in
the publication. For instance, \code{"catch.csv"} or \code{c("catch.csv", "trap.csv")}}

\item{attributes_file_names}{A list of file names for the attributes metadata
csv templates. The naming convention for the files should be \code{attributes_NAME_OF_DATA_TABLE.csv}}

\item{title}{Title of the data publication. The title should be clear, concise,
and descriptive.}

\item{maintenance}{This describes how often the data publication will be updated.
The options are: \code{daily}, \code{weekly}, \code{monthly}, \code{annually}, \code{complete}.}

\item{edi_number}{The EDI number to be used for the data publication. If the EML
is generated to update an existing package, the number for the next version should be the input.
For instance, if the existing EDI package is \code{edi.1234.1} then the you should
input \code{edi.1234.2} because you are creating the EML for the next version.
If this is for a new package, please reserve a new EDI number. Refer to the
\href{https://lucy-dwr.github.io/hrl-docs/quickstart/publish-to-edi.html}{Quickstart Guide}.}
}
\value{
An EML file within \code{data-raw/eml}. EMLassemblyline will include messages
to ensure the EML file validates and will also flag any issues.
}
\description{
This is a wrapper function and utilizes the package \href{https://ediorg.github.io/EMLassemblyline/}{EMLassemblyline} to
facilitate the creation of an EML document
}
\details{
This function assumes that all appropriate metadata templates have been completed
Check that the following files in \code{publish/metadata_templates} have been manually filled
in and are correct and generate any additional files as needed:
abstract.txt
attributes_csv_templates/attributes_csv_template.csv (one per data table)
custom_units.txt
geographic_coverage.txt
intellectual_rights.txt (DO NOT CHANGE unless using another license)
keywords.txt
methods.docx
personnel.txt

This function also assumes that the user has an EDI account and credentials
are stored in their .Renviron as EDI_USER_ID
}
\examples{
\dontrun{
make_eml_edi(data_file_names = c("microhabitat_observations.csv","survey_locations.csv"),
             attributes_file_names = c("attributes_microhabitat_observations.csv","attributes_survey_locations.csv"),
             title = "Distribution and habitat use of juvenile steelhead and other fishes of the lower Feather River",
             maintenance = "complete",
             edi_number = "edi.test")
}

}
