% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-qc-checks.R
\name{print_qc_summary}
\alias{print_qc_summary}
\title{Print QC Summary}
\usage{
print_qc_summary(qc_results)
}
\arguments{
\item{qc_results}{List output from run_qc_checks()}
}
\description{
Print QC Summary
}
