% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish-data-edi.R
\name{publish_data_edi}
\alias{publish_data_edi}
\title{Publish data package on EDI}
\usage{
publish_data_edi(path_eml_file, publish_type, edi_number, publish_environment)
}
\arguments{
\item{publish_type}{Indicate whether you are publishing an update to an existing
package or publishing a new data package. Option are: "new", "update"}

\item{edi_number}{EDI number of the data package being published. This should
match the EDI number in the EML document. If you are updating an existing package
the EDI number will be one version greater than the existing package.}

\item{publish_environment}{Environment where data are being published. Recommended to
publish on the staging environment first to review the data package and then
publish to production. Option are: "staging", "production"}
}
\value{
A message describing if the upload was successful or not.
}
\description{
This is a wrapper function and utilizes the package EMLaide to send an API
request to EDI.
}
\details{
This function assumes that `make_eml_edi` was used to create an EML metadata
document and that the file structure follows the documentation. INSERT link.

This function also assumes that the user has an EDI account and credentials
are stored in their .Renviron as EDI_USER_ID and EDI_PASSWORD
}
