% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-qc-checks.R
\name{qc_carcass_survey}
\alias{qc_carcass_survey}
\title{QC Checks for Carcass Survey Data}
\usage{
qc_carcass_survey(data)
}
\arguments{
\item{data}{Data frame with columns: date, site_id, survey_period,
carcass_id, condition, sex, fork_length, marked}
}
\value{
List with flags and summary
}
\description{
QC Checks for Carcass Survey Data
}
