% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-qc-checks.R
\name{qc_habitat}
\alias{qc_habitat}
\title{QC Checks for Habitat Measurement Data}
\usage{
qc_habitat(data)
}
\arguments{
\item{data}{Data frame with columns: date, site_id, transect_id,
depth, velocity, substrate (optional)}
}
\value{
List with flags and summary
}
\description{
QC Checks for Habitat Measurement Data
}
