% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-qc-checks.R
\name{qc_invertebrate}
\alias{qc_invertebrate}
\title{QC Checks for Invertebrate Sample Data}
\usage{
qc_invertebrate(data)
}
\arguments{
\item{data}{Data frame with columns: date, site_id, sample_id, taxon, count}
}
\value{
List with flags and summary
}
\description{
QC Checks for Invertebrate Sample Data
}
